musl-libc 0/4, extract au_nftw() fot two versions
In order to support both of glibc and musl-libc, an internal function
au_nftw() was introduced by
	c03368e 2016-07-27 auplink_ftw to support non-glibc
But more work is necessary to fully support musl. This commit is the
first one of this series.

This commit does
- for au_ntfw(), extract it from plink.c and create two versions under
  lib2 dir. One is for glibc and the other is for musl.
- a new internal make var Glibc replaces NoLibcFTW which was given by a
  user, and new Glibc is the switch to decide which version of au_nftw()
  to use.

The behaviour doesn't change.
